LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-choose-value-from-picker-expected.txt

Tests that choosing a value using the date/time picker does not unexpectedly add or remove second and millisecond fields.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing empty input with no step value.
PASS valueHasSeconds is false
PASS valueHasMillseconds is false

Testing empty input with step attribute precise to seconds.
PASS valueHasSeconds is true
PASS valueHasMillseconds is false

Testing empty input with step attribute precise to milliseconds.
PASS valueHasSeconds is true
PASS valueHasMillseconds is true

Testing input with a value that contains seconds.
PASS valueHasSeconds is true
PASS valueHasMillseconds is false

Testing input with a value that contains milliseconds.
PASS valueHasSeconds is true
PASS valueHasMillseconds is true

PASS successfullyParsed is true

TEST COMPLETE
